Separate movieInfo to it's own component
