+++ frontend/app/movies/[id]/page.tsx
@@ -1,40 +1,19 @@
 "use client";
 
-import { Dropdown } from "@/components/dropdown";
-import { StarRating } from "@/components/starRating";
+import { MovieInfo } from "@/components/movies/movieInfo";
+import { Section } from "@/components/section";
 
 import { type MovieCrew, type Movies } from "@prisma/client";
 import { notFound } from "next/navigation";
 
 import { useEffect, useState } from "react";
-import { Heart, Star } from "react-feather";
+import { Star } from "react-feather";
 
 export default function Movie({ params }: { params: { id: string } }) {
-  const [movie, setMovie] = useState<Movies | null>(null);
-  const [directors, setDirectors] = useState<string[]>([]);
-
-  const [watched, setWatched] = useState<boolean>(false);
-  const [favorite, setFavorite] = useState<boolean>(false);
-  const [watchlist, setWatchlist] = useState<boolean>(false);
-
-  const [userRating, setUserRating] = useState<number>(0);
-
   const [isLoading, setIsLoading] = useState<boolean>(true);
 
-  const handleRatingChange = (rating: number) => {
-    setUserRating(rating);
-  };
-
-  const placeholderIcon = {
-    backgroundColor: "rgba(0,0,0,0.25)",
-    color: "rgba(255,255,255,0.35)",
-    padding: "5px",
-    aspectRatio: 1,
-    display: "flex",
-    alignItems: "center",
-    fontSize: "0.75rem",
-    borderRadius: "50%",
-  };
+  const [movie, setMovie] = useState<Movies | null>(null);
+  const [directors, setDirectors] = useState<string[]>([]);
 
   useEffect(() => {
     const getMovie = async () => {
@@ -55,20 +34,6 @@ export default function Movie({ params }: { params: { id: string } }) {
     getMovie();
   }, [params.id]);
 
-  if (isLoading) {
-    return (
-      <main className="rotating">
-        <Star />
-      </main>
-    );
-  } else if (!isLoading && !movie) {
-    notFound();
-  }
-
-  if (!movie) {
-    return null;
-  }
-
   async function getDirectors(crewArray: MovieCrew[]) {
     let directorIds: number[] = [];
     let directorNames: string[] = [];
@@ -85,83 +50,47 @@ export default function Movie({ params }: { params: { id: string } }) {
     setIsLoading(false);
   }
 
+  if (isLoading) {
+    return (
+      <main className="rotating">
+        <Star />
+      </main>
+    );
+  } else if (!isLoading && !movie) {
+    notFound();
+  }
+
+  if (!movie) {
+    notFound();
+  }
+
   return (
     <main style={{ padding: 0 }}>
-      <div className="movie-content">
-        <div className="wrapper">
-          {/* Img not working until we have posters to path to, can be tested with any img url until then */}
-          <div
-            className="img"
-            style={{
-              background: `URL(${movie.poster_path})`,
-              backgroundPosition: "center",
-              backgroundSize: "cover",
-            }}
-          >
-            <div className="gradient" />
-          </div>
+      <MovieInfo
+        movie={movie}
+        directors={directors}
+      />
 
-          <div className="info">
-            <div className="rating">
-              <div className="current-rating">{movie.vote_average ? movie.vote_average.toFixed(1) : null}</div>
-              <StarRating onChange={handleRatingChange} />
-            </div>
-            <div className="basic-data">
-              <h1>{movie.title}</h1>
+      <div className="section-padding">
+        <Section header="Cast">
+          <p>Coming soon</p>
+        </Section>
 
-              <div className="data-row">
-                {directors.length > 0 ? <p className="yellow">Directed by {directors.join(", ")}</p> : null}
-                <p className="cyan">
-                  {movie.release_date ? new Date(movie.release_date as Date).getFullYear().toString() : null}
-                </p>
-                <p className="cyan">(Age restriction)</p>
-                <p className="cyan">{movie.runtime ? movie.runtime : null} min</p>
-              </div>
-              <p className="movie-description">{movie.overview}</p>
-            </div>
+        <Section header="Crew">
+          <p>Coming soon</p>
+        </Section>
 
-            <div className="all-buttons">
-              <div className="buttons">
-                <Dropdown
-                  button={<button>Add to list</button>}
-                  width={200}
-                >
-                  <button className="dropdown-item">Example list 1</button>
-                  <button className="dropdown-item">Example list 2</button>
-                  <button className="dropdown-item">Example list 3</button>
-                </Dropdown>
-                <button
-                  className={watched ? "button-pink" : ""}
-                  onClick={() => setWatched(!watched)}
-                >
-                  {watched ? "Remove from watched" : "Mark as watched"}
-                </button>
-              </div>
-              <div className="transparent-buttons">
-                <button
-                  className={favorite ? "button-transparent pink" : "button-transparent"}
-                  onClick={() => setFavorite(!favorite)}
-                >
-                  <Heart size={24} />
-                  {favorite ? "Remove from favorites" : "Add to favorites"}
-                </button>
-                <button
-                  className={watchlist ? "button-transparent yellow" : "button-transparent"}
-                  onClick={() => setWatchlist(!watchlist)}
-                >
-                  <Star size={24} />
-                  {watchlist ? "Remove from watchlist" : "Add to watchlist"}
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div className="websites">
-          <h6>Watch at:</h6>
-          <p style={placeholderIcon}>icon</p>
-          <p style={placeholderIcon}>icon</p>
-          <p style={placeholderIcon}>icon</p>
-        </div>
+        <Section header="Reviews">
+          <p>Coming soon</p>
+        </Section>
+
+        <Section header="In theaters">
+          <p>Coming soon</p>
+        </Section>
+
+        <Section header="Similar movies">
+          <p>Coming soon</p>
+        </Section>
       </div>
     </main>
   );
