+++ frontend/sass/pages/_movies.scss
@@ -1,6 +1,10 @@
 @use "sass:map";
 @use "../abstracts" as abstracts;
 
+.section-padding {
+  padding: 60px 60px 40px 60px;
+}
+
 .movie-content {
   background-color: abstracts.$primary-grey;
   box-shadow: abstracts.$default-shadow;
@@ -170,6 +174,9 @@
         flex-direction: column;
         flex-wrap: wrap;
         gap: 10px;
+      }
+      .section-padding {
+        padding: 0;
       }
     }
   }
